M4: Create runtime_checks.py with V family and regime-prob family
Create `regime_building/runtime_checks.py` and absorb two families
from `utils/error_handling.py`:
- V family (`validate_V`, `_enrich_with_diagnostics`,
  `_summarize_diagnostics`, `_format_diagnostic_summary`)
- regime-prob family (`validate_regime_transition_probs` ->
  `_validate_regime_transition_probs`, `_format_sum_violation`,
  `validate_regime_transitions_all_periods`,
  `_validate_regime_transition_single`,
  `_validate_no_reachable_incomplete_targets`)

Both families fit the unifying concept "defensive checks on JAX
arrays produced during solve/simulate." Privatize
`validate_regime_transition_probs` (only tests call it directly).
Update imports in 6 callers (3 src, 3 test). `diagnostics.py` keeps
its name. Part of Phase 1 — see `Phase 1 — Validation Cleanup.md`.
