+++ src/lcm/solution/solve_brute.py
@@ -12,8 +12,8 @@ import jax.numpy as jnp
 
 from lcm.ages import AgeGrid
 from lcm.interfaces import Regime, _build_regime_sharding
+from lcm.regime_building.runtime_checks import validate_V
 from lcm.typing import BoolND, FlatParams, FloatND, RegimeName, StateName
-from lcm.utils.error_handling import validate_V
 from lcm.utils.logging import (
     format_duration,
     log_period_header,
