+++ tests/test_nan_diagnostics.py
@@ -10,6 +10,7 @@ from lcm.ages import AgeGrid
 from lcm.exceptions import InvalidValueFunctionError
 from lcm.grids import LinSpacedGrid
 from lcm.interfaces import StateActionSpace
+from lcm.regime_building.runtime_checks import validate_V
 from lcm.typing import (
     BoolND,
     ContinuousAction,
@@ -18,7 +19,6 @@ from lcm.typing import (
     ScalarInt,
 )
 from lcm.user_regime import Regime as UserRegime
-from lcm.utils.error_handling import validate_V
 
 
 def _make_state_action_space(
