+++ tests/test_validate_regime_transition_probs.py
@@ -12,12 +12,12 @@ from lcm import (
     categorical,
 )
 from lcm.exceptions import InvalidRegimeTransitionProbabilitiesError
+from lcm.regime_building.runtime_checks import (
+    _format_sum_violation,
+    _validate_regime_transition_probs,
+)
 from lcm.typing import DiscreteAction, FloatND, ScalarInt
 from lcm.user_regime import Regime as UserRegime
-from lcm.utils.error_handling import (
-    _format_sum_violation,
-    validate_regime_transition_probs,
-)
 from lcm_examples.mortality import RegimeId as MortalityRegimeId
 from lcm_examples.mortality import get_model, get_params
 
@@ -30,7 +30,7 @@ def test_valid_probs_all_active():
             "retirement": jnp.array([0.3, 0.4]),
         }
     )
-    validate_regime_transition_probs(
+    _validate_regime_transition_probs(
         regime_transition_probs=probs,
         active_regimes_next_period=("working_life", "retirement"),
         regime_name="working_life",
@@ -48,7 +48,7 @@ def test_valid_probs_with_inactive_regime_at_zero():
             "dead": jnp.array([0.0, 0.0]),
         }
     )
-    validate_regime_transition_probs(
+    _validate_regime_transition_probs(
         regime_transition_probs=probs,
         active_regimes_next_period=("working_life", "retirement"),
         regime_name="working_life",
@@ -69,7 +69,7 @@ def test_raises_for_probs_not_summing_to_one():
         InvalidRegimeTransitionProbabilitiesError,
         match=r"2 of 3 probability vectors do not sum to 1\.0",
     ):
-        validate_regime_transition_probs(
+        _validate_regime_transition_probs(
             regime_transition_probs=probs,
             active_regimes_next_period=("working_life", "retirement"),
             regime_name="working_life",
@@ -91,7 +91,7 @@ def test_raises_for_positive_probability_on_inactive_regime():
         InvalidRegimeTransitionProbabilitiesError,
         match=r"'dead' is inactive at age 26\.0",
     ):
-        validate_regime_transition_probs(
+        _validate_regime_transition_probs(
             regime_transition_probs=probs,
             active_regimes_next_period=("working_life", "retirement"),
             regime_name="working_life",
@@ -112,7 +112,7 @@ def test_raises_for_out_of_bounds_values():
         InvalidRegimeTransitionProbabilitiesError,
         match=r"values outside \[0, 1\]",
     ):
-        validate_regime_transition_probs(
+        _validate_regime_transition_probs(
             regime_transition_probs=probs,
             active_regimes_next_period=("working_life", "retirement"),
             regime_name="working_life",
@@ -133,7 +133,7 @@ def test_raises_for_nan_values():
         InvalidRegimeTransitionProbabilitiesError,
         match=r"Non-finite values.*between ages 25\.0 and 26\.0",
     ):
-        validate_regime_transition_probs(
+        _validate_regime_transition_probs(
             regime_transition_probs=probs,
             active_regimes_next_period=("working_life", "retirement"),
             regime_name="working_life",
@@ -154,7 +154,7 @@ def test_raises_for_inf_values():
         InvalidRegimeTransitionProbabilitiesError,
         match="Non-finite values",
     ):
-        validate_regime_transition_probs(
+        _validate_regime_transition_probs(
             regime_transition_probs=probs,
             active_regimes_next_period=("working_life", "retirement"),
             regime_name="working_life",
